Reorganize the code, fix typos, and update visibility of internal types
I thought it would be better to move all the GitHub search related code
into its own module. At the same time I thought it would be a good idea
to fix some typos that I saw and restrict access to types that are only
used internally.
