+++ src/github/graphql.rs
@@ -4,9 +4,9 @@ use time::OffsetDateTime;
 
 /// Details on the endpoint can be found here
 /// <https://docs.github.com/en/graphql/guides/forming-calls-with-graphql#the-graphql-endpoint>
-pub(crate) const GITHUB_GRAPHQL_URL: &str = "https://api.github.com/graphql";
+pub(super) const GITHUB_GRAPHQL_URL: &str = "https://api.github.com/graphql";
 
-pub(crate) const GITHUB_REPOSITORY_QUERY: &str = "
+pub(super) const GITHUB_REPOSITORY_QUERY: &str = "
 query GitHubRepositorySearch(
   # The search string to look for. GitHub search syntax is supported.
   $gitHubSearchString: String!
@@ -56,7 +56,7 @@ query GitHubRepositorySearch(
 }
 ";
 
-pub(crate) fn github_repository_search_variables(
+pub(super) fn github_repository_search_variables(
     limit: usize,
     cursor_offset: Option<&str>,
     min_stars: Option<usize>,
@@ -74,7 +74,7 @@ pub(crate) fn github_repository_search_variables(
 }
 
 #[derive(Debug, Deserialize)]
-pub struct GraphQLResponse<T> {
+pub(super) struct GraphQLResponse<T> {
     pub data: Option<T>,
     pub error: Option<serde_json::Value>,
 }
@@ -83,7 +83,7 @@ impl<T> GraphQLResponse<T>
 where
     T: DeserializeOwned,
 {
-    pub fn new(source: String) -> Result<Self, serde_json::Error> {
+    pub(super) fn new(source: String) -> Result<Self, serde_json::Error> {
         serde_json::from_str(&source)
     }
 }
@@ -102,13 +102,13 @@ struct PageInfo {
 
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "camelCase")]
-pub struct GitHubSearchResult {
+pub(super) struct GitHubSearchResult {
     search: GitHubSearchResultInner,
 }
 
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "camelCase")]
-pub struct GitHubSearchResultInner {
+struct GitHubSearchResultInner {
     /// The total number of repositories that matched the search query
     #[allow(dead_code)]
     repository_count: usize,
@@ -121,57 +121,57 @@ pub struct GitHubSearchResultInner {
 
 impl GitHubSearchResult {
     /// The total number of repositories that met the search critera.
-    #[allow(dead_code)]
-    pub fn total_repository_count(&self) -> usize {
+    pub(super) fn total_repository_count(&self) -> usize {
         self.search.repository_count
     }
 
     /// All the repositories returned in this page of data.
     #[allow(dead_code)]
-    pub fn repositories(&self) -> &[RepositoryInfo] {
+    pub(super) fn repositories(&self) -> &[RepositoryInfo] {
         &self.search.repositories
     }
 
     /// Convert the search result into a [Vec<RepositoryInfo>].
-    pub fn into_repositories(self) -> Vec<RepositoryInfo> {
+    pub(super) fn into_repositories(self) -> Vec<RepositoryInfo> {
         self.search.repositories
     }
 
     /// Token for the next page of data if it exists.
-    pub fn next_page(&self) -> Option<&str> {
+    pub(super) fn next_page(&self) -> Option<&str> {
         self.search.page_info.end_cursor.as_deref()
     }
 }
 
+/// Metadata about this GitHub Repository
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct RepositoryInfo {
     /// The Node ID of the Repository object.
-    /// The [ID] represents a unique identifier that is Base64 obfuscated
+    /// The [`ID`] represents a unique identifier that is Base64 obfuscated
     ///
-    /// [ID]: https://docs.github.com/en/graphql/reference/scalars#id
+    /// [`ID`]: https://docs.github.com/en/graphql/reference/scalars#id
     id: String,
     /// The repository's name with owner.
     name_with_owner: String,
     /// The description of the repository.
     // FIXME(ytmimi) I eventualy plan to write this value to the database,
-    // but I don't need it no.
+    // but I don't need it now.
     #[allow(unused)]
     description: String,
     /// Identifies the date and time when the repository was archived.
     #[serde(with = "time::serde::iso8601::option")]
     // FIXME(ytmimi) I eventualy plan to write this value to the database,
-    // but I don't need it no.
+    // but I don't need it now.
     #[allow(unused)]
     archived_at: Option<OffsetDateTime>,
     /// Identifies if the repository is a fork.
     // FIXME(ytmimi) I eventualy plan to write this value to the database,
-    // but I don't need it no.
+    // but I don't need it now.
     #[allow(unused)]
     is_fork: bool,
     /// Indicates if the repository has been locked or not.
     // FIXME(ytmimi) I eventualy plan to write this value to the database,
-    // but I don't need it no.
+    // but I don't need it now.
     #[allow(unused)]
     is_locked: bool,
     /// Identifies the date and time when the repository was last pushed to.
@@ -187,25 +187,29 @@ pub struct RepositoryInfo {
 }
 
 impl RepositoryInfo {
-    /// Get the GitHub GraphQL ID for this repository
+    /// Get the GitHub GraphQL ID for this repository.
+    /// The [ID] represents a unique identifier that is Base64 obfuscated
+    ///
+    /// [ID]: https://docs.github.com/en/graphql/reference/scalars#id
     pub fn id(&self) -> &str {
         self.id.as_str()
     }
 
     /// Get the name of the repostor with the owner included.
+    /// For example `rust-lang/rust`
     pub fn name_with_owner(&self) -> &str {
         self.name_with_owner.as_str()
     }
 
-    /// Provides an iterator over the languages in this git repository
-    pub fn languages<'a>(&'a self) -> LanguageIterator<impl Iterator<Item = &'a LanguageNode>> {
+    /// Provides an iterator over the languages in this git repository.
+    pub fn languages(&self) -> impl Iterator<Item = ProgrammingLanguage<'_>> {
         LanguageIterator {
             total_size: self.languages.total_size,
             inner: self.languages.edges.iter(),
         }
     }
 
-    /// How much of this repository was written in Rust
+    /// How much of this repository was written in Rust.
     pub fn percent_of_code_in_rust(&self) -> f64 {
         self.languages()
             .filter(|programming_language| programming_language.name() == "Rust")
@@ -215,17 +219,17 @@ impl RepositoryInfo {
             })
     }
 
-    /// Returns a reference to the latest commit hash fetched from GitHub
+    /// Returns a reference to the latest commit hash fetched from GitHub.
     pub fn commit_hash(&self) -> &str {
         &self.default_branch_ref.target.oid
     }
 
-    /// The last time that the repository was pushe
+    /// The last time this repository was pushed.
     pub fn pushed_at(&self) -> OffsetDateTime {
         self.pushed_at
     }
 
-    /// The last time that the repository was pushe
+    /// The last time this repository was updated.
     pub fn updated_at(&self) -> OffsetDateTime {
         self.updated_at
     }
@@ -244,17 +248,18 @@ struct Languages {
     edges: Vec<LanguageNode>,
 }
 
-pub struct LanguageIterator<I> {
+struct LanguageIterator<I> {
     total_size: usize,
     inner: I,
 }
 
-pub struct ProgramingLanguage<'a> {
+/// Programming language used by a GitHub repository.
+pub struct ProgrammingLanguage<'a> {
     percent_of_code_in_repo: f64,
     name: &'a str,
 }
 
-impl<'a> ProgramingLanguage<'a> {
+impl<'a> ProgrammingLanguage<'a> {
     pub fn name(&self) -> &str {
         self.name
     }
@@ -268,10 +273,10 @@ impl<'a, I> Iterator for LanguageIterator<I>
 where
     I: Iterator<Item = &'a LanguageNode>,
 {
-    type Item = ProgramingLanguage<'a>;
+    type Item = ProgrammingLanguage<'a>;
 
     fn next(&mut self) -> Option<Self::Item> {
-        self.inner.next().map(|language_node| ProgramingLanguage {
+        self.inner.next().map(|language_node| ProgrammingLanguage {
             percent_of_code_in_repo: ((language_node.size as f64) / (self.total_size as f64))
                 * 100f64,
             name: &language_node.node.name,
