+++ src/lib.rs
@@ -1,3 +1,3 @@
 mod github;
 
-pub use github::GitHubRepoSearch;
+pub use github::{GitHubRepoSearch, ProgrammingLanguage, RepoSearchResults, RepositoryInfo};
